assistant_app/requirements.txt
PySide6>=6.7
python-dotenv>=1.0
httpx>=0.27
requests>=2.32
numpy>=1.26
sounddevice>=0.4
faster-whisper>=1.0
opencv-python>=4.10
face_recognition>=1.3
Pillow>=10.4
pytest>=8.3
pytest-qt>=4.4
